\name{strip.slashes}
\alias{stirp.slashes}
\title{Remove slashes from inside a string}
\description{
This function removes all slashes inside all elements of a character vector.
}

\usage{
strip.slashes(char.vector)
}

\arguments{
  \item{char.vector}{A vector of characters}
}


\details{
There really isn't much more to it than that.
}

\value{
A vector of equal length as \code{from} with all internal slashes
removed.
}

\author{Steve Lianoglou}

% \seealso{
%   \code{\link{strip.whitespace}}
% }
\examples{
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{utilities}
